move new functions into src, adjust module syntax, add to index.ts
