--- index.ts
@@ -1,7 +1,0 @@
-module.exports.arrayFromRange = require( './arrayFromRange' );
-module.exports.hexToRGB = require( './hexToRgb' );
-module.exports.minifyCss = require( './minifyCss' );
-module.exports.parseDateGMT = require( './parseDateGMT' );
-module.exports.resizeWPImage = require( './resizeWPImage' );
-module.exports.stylizedTimestamp = require( './stylizedTimestamp' );
-module.exports.timeAgo = require( './timeAgo' );
