+++ src/index.ts
@@ -1,4 +1,7 @@
 export { default as arrayFromRange } from './arrayFromRange';
 export { default as hexToRGB } from './hexToRGB';
 export { default as minifyCss } from './minifyCss';
+export { default as parseDateGMT } from './parseDateGMT';
 export { default as resizeWPImage } from './resizeWPImage';
+export { default as stylizedTimestamp } from './stylizedTimestamp';
+export { default as timeAgo } from './timeAgo';
